[ALPS02364400] sensors: 6797 add virtual sensors
[Detail] 6797 add virtual sensors

[Solution] 6797 add kernel virtual sensor code

[Feature] Others

MTK-Commit-Id: 6cf72299191fe832dffff96cd9233bcf981b52a0

CR-Id: ALPS02364400

+++ drivers/misc/mediatek/glance_gesture_sensor/glance_gesture.h
@@ -51,7 +51,7 @@ struct glg_control_path {
 };
 
 struct glg_data_path {
-	int (*get_data)(u16 *value, int *status);
+	int (*get_data)(int *value, int *status);
 };
 
 struct glg_init_info {
